ref: move getActions to own file
